+++ framework/db-core/src/test/java/is/codion/framework/db/TestDomain.java
@@ -136,7 +136,7 @@ public final class TestDomain extends DomainModel {
 										Employee.DEPARTMENT_LOCATION.define()
 														.denormalized()
 														.from(Employee.DEPARTMENT_FK)
-														.attribute(Department.LOCATION)
+														.using(Department.LOCATION)
 														.caption(Department.LOCATION.name()))
 						.table("employees.employee")
 						.orderBy(ascending(Employee.DEPARTMENT, Employee.NAME))