+++ framework/domain/src/test/java/is/codion/framework/domain/TestDomain.java
@@ -248,7 +248,7 @@ public final class TestDomain extends DomainModel {
 										Detail2.MASTER_NAME.define()
 														.denormalized()
 														.from(Detail2.MASTER_FK)
-														.attribute(Master2.NAME)
+														.using(Master2.NAME)
 														.caption(Detail2.MASTER_NAME.name()),
 										Detail2.MASTER_CODE.define()
 														.column()
@@ -359,12 +359,12 @@ public final class TestDomain extends DomainModel {
 										Detail.MASTER_NAME.define()
 														.denormalized()
 														.from(Detail.MASTER_FK)
-														.attribute(Master.NAME)
+														.using(Master.NAME)
 														.caption(Detail.MASTER_NAME.name()),
 										Detail.MASTER_CODE.define()
 														.denormalized()
 														.from(Detail.MASTER_FK)
-														.attribute(Master.CODE)
+														.using(Master.CODE)
 														.caption(Detail.MASTER_CODE.name()),
 										Detail.INT_ITEMS.define()
 														.column()
@@ -520,7 +520,7 @@ public final class TestDomain extends DomainModel {
 										Employee.DEPARTMENT_LOCATION.define()
 														.denormalized()
 														.from(Employee.DEPARTMENT_FK)
-														.attribute(Department.LOCATION)
+														.using(Department.LOCATION)
 														.caption(Department.LOCATION.name()),
 										Employee.DEPARTMENT_NAME.define()
 														.derived()