+++ framework/servlet/src/test/java/is/codion/framework/servlet/TestDomain.java
@@ -146,7 +146,7 @@ public final class TestDomain extends DomainModel {
 										Employee.DEPARTMENT_LOCATION.define()
 														.denormalized()
 														.from(Employee.DEPARTMENT_FK)
-														.attribute(Department.LOCATION)
+														.using(Department.LOCATION)
 														.caption(Department.LOCATION.name()),
 										Employee.DATA.define()
 														.column()